+++ examples/encode_raw_rgb.c
@@ -48,7 +48,7 @@ static int encode(int width, int height, const char *input_filename, struct enco
         }
 
         // load image and set it as the encoder input buffer
-        int input_image_size = 0;
+        size_t input_image_size = 0;
         if (gpujpeg_image_load_from_file(input_filename, &d->input_image, &input_image_size) != 0) {
                 return 1;
         }
@@ -57,7 +57,7 @@ static int encode(int width, int height, const char *input_filename, struct enco
 
         // compress the image
         uint8_t *image_compressed = NULL;
-        int image_compressed_size = 0;
+        size_t image_compressed_size = 0;
         if (gpujpeg_encoder_encode(d->encoder, &param, &param_image, &encoder_input, &image_compressed, &image_compressed_size) != 0) {
                 return 1;
         }
